Decouple the scenario config from HttpMock
