+++ integration-tests/oauth-mock-test-runner/src/main/groovy/com/okta/test/mock/scenarios/ImplicitRemoteValidationScenarioDefinition.groovy
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2017 Okta, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.okta.test.mock.scenarios
+
+import com.github.tomakehurst.wiremock.WireMockServer
+
+import static com.github.tomakehurst.wiremock.client.WireMock.aResponse
+import static com.github.tomakehurst.wiremock.client.WireMock.containing
+import static com.github.tomakehurst.wiremock.client.WireMock.get
+import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo
+
+class ImplicitRemoteValidationScenarioDefinition implements ScenarioDefinition {
+
+    void configureHttpMock(WireMockServer wireMockServer) {
+        wireMockServer.stubFor(
+                get("/oauth2/default/.well-known/openid-configuration")
+                    .willReturn(aResponse()
+                        .withHeader("Content-Type", "application/json")
+                        .withBodyFile("discovery.json")
+                        .withTransformers("gstring-template")))
+
+        wireMockServer.stubFor(
+                get(urlPathEqualTo("/oauth2/default/v1/keys"))
+                    .willReturn(aResponse()
+                        .withHeader("Content-Type", "application/json;charset=UTF-8")
+                        .withBodyFile("keys.json")
+                        .withTransformers("gstring-template")))
+
+        wireMockServer.stubFor(
+                get(urlPathEqualTo("/oauth2/default/v1/userinfo"))
+                    .withHeader("Authorization", containing("Bearer some.random.jwt"))
+                        .willReturn(aResponse()
+                        .withHeader("Content-Type", "application/json;charset=UTF-8")
+                        .withBodyFile("userinfo-remote-access-token.json")))
+    }
+}
