+++ integration-tests/oauth-mock-test-runner/src/main/groovy/com/okta/test/mock/scenarios/ScenarioDefinition.groovy
@@ -13,13 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.okta.test.mock
+package com.okta.test.mock.scenarios
 
-class OAuthMockServer {
+import com.github.tomakehurst.wiremock.WireMockServer
 
-    static main(String[] args) {
+trait ScenarioDefinition {
 
-
-
+    Map getBindingMap() {
+        return Collections.emptyMap()
     }
+
+    abstract void configureHttpMock(WireMockServer wireMockServer)
 }
